+++ shared/src/commonMain/kotlin/com/dz/tmdbmobileapi/usecase/GetPopularMoviesUseCase.kt
@@ -1,6 +1,6 @@
 package com.dz.tmdbmobileapi.usecase
 
-import com.dz.tmdbmobileapi.remote.PopularMoviesDataSource
+import com.dz.tmdbmobileapi.remote.datasource.PopularMoviesDataSource
 
 class GetPopularMoviesUseCase(private val popularMoviesDataSource: PopularMoviesDataSource) {
     suspend operator fun invoke() = popularMoviesDataSource.getPopularMovies(1)
